# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Job Search Platform API",
        "description": "API Documentation for Job Search Platform Project",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/courses/{id}": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "Get course details",
                "description": "Returns detailed information about a specific course.",
                "operationId": "getCourseDetails",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Course details"
                    }
                }
            }
        },
        "/courses/{id}/enroll": {
            "post": {
                "tags": [
                    "Courses"
                ],
                "summary": "Enroll in a course",
                "description": "Enroll in a course.",
                "operationId": "enrollCourse",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Enrolled successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Courses"
                ],
                "summary": "Unenroll from a course",
                "description": "Cancel enrollment (unenroll).",
                "operationId": "unenrollCourse",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Unenrolled successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/courses/my-enrollments": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "Get my enrollments",
                "description": "Get my enrollments.",
                "operationId": "getMyEnrollments",
                "responses": {
                    "200": {
                        "description": "List of enrollments"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/employer/courses": {
            "get": {
                "tags": [
                    "Employer",
                    "Courses"
                ],
                "summary": "Get employer's courses",
                "description": "Get employer's courses.",
                "operationId": "getEmployerCourses",
                "responses": {
                    "200": {
                        "description": "List of employer's courses"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Employer",
                    "Courses"
                ],
                "summary": "Create a new course",
                "description": "Create a new course ad.",
                "operationId": "createCourse",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "topics": {
                                        "type": "string"
                                    },
                                    "duration": {
                                        "type": "string"
                                    },
                                    "location": {
                                        "type": "string"
                                    },
                                    "trainer": {
                                        "type": "string"
                                    },
                                    "fees": {
                                        "type": "number"
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "Draft",
                                            "Active"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Course created successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/employer/courses/{id}": {
            "put": {
                "tags": [
                    "Employer",
                    "Courses"
                ],
                "summary": "Update course details",
                "description": "Update a course ad.",
                "operationId": "updateCourse",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "topics": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Course updated successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Employer",
                    "Courses"
                ],
                "summary": "Delete a course",
                "description": "Delete a course ad.",
                "operationId": "deleteCourse",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Course deleted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/employer/courses/{id}/publish": {
            "post": {
                "tags": [
                    "Employer",
                    "Courses"
                ],
                "summary": "Publish a course",
                "description": "Publish a course (change status to Active).",
                "operationId": "publishCourse",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Course published"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/employer/courses/{id}/close": {
            "post": {
                "tags": [
                    "Employer",
                    "Courses"
                ],
                "summary": "Close a course",
                "description": "Close a course.",
                "operationId": "closeCourse",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Course closed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/employer/courses/{id}/enrollments": {
            "get": {
                "tags": [
                    "Employer",
                    "Courses"
                ],
                "summary": "Get enrollments for a course",
                "description": "Get course enrollments (for employer).",
                "operationId": "getCourseEnrollments",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of enrollments"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}